Dump a SummarizedExperiment to disk

Dump a SummarizedExperiment (or one of its subclasses) to disk in the wobbegong format for easy HTTP range requests.

wobbegongify(x, path, dense.assays = TRUE, sparse.assays = TRUE)

Arguments

x

A SummarizedExperiment or an instance of one of its subclasses.

path

String containing a path to the directory to dump x.

dense.assays

Logical scalar indicating whether to save dense assays. If FALSE, dense assays are skipped.

sparse.assays

Logical scalar indicating whether to save dense assays. If FALSE, sparse assays are skipped.

Value

path is populated with the contents of x. NULL is returned invisibly.

Author

Aaron Lun

Examples

library(SingleCellExperiment)
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> Loading required package: GenomicRanges
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> Loading required package: generics
#> 
#> Attaching package: ‘generics’
#> The following objects are masked from ‘package:base’:
#> 
#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
#>     setequal, union
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, aperm, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, is.unsorted, lapply,
#>     mapply, match, mget, order, paste, pmax, pmax.int, pmin, pmin.int,
#>     rank, rbind, rownames, sapply, saveRDS, table, tapply, unique,
#>     unsplit, which.max, which.min
#> Loading required package: S4Vectors
#> 
#> Attaching package: ‘S4Vectors’
#> The following object is masked from ‘package:utils’:
#> 
#>     findMatches
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: IRanges
#> Loading required package: GenomeInfoDb
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
#> 
#> Attaching package: ‘Biobase’
#> The following object is masked from ‘package:MatrixGenerics’:
#> 
#>     rowMedians
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     anyMissing, rowMedians
se <- SingleCellExperiment(
    assays = list(
        counts=matrix(rpois(200, lambda=5), ncol=10), 
        logcounts=matrix(rnorm(200), ncol=10)
    ),
    colData = DataFrame(
        yy = letters[1:10],
        xx = LETTERS[1:10]
    ),
    rowData = DataFrame(row.names=sprintf("GENE_%i", 1:20)),
    reducedDims = list(
        PCA = matrix(rnorm(50), nrow=10),
        TSNE = matrix(rnorm(20), nrow=10)
    )
)

tmp <- tempfile()
wobbegongify(se, tmp)
list.files(tmp, recursive=TRUE)
#>  [1] "assays/0/content"                  "assays/0/stats"                   
#>  [3] "assays/0/summary.json"             "assays/1/content"                 
#>  [5] "assays/1/stats"                    "assays/1/summary.json"            
#>  [7] "column_data/content"               "column_data/summary.json"         
#>  [9] "reduced_dimensions/0/content"      "reduced_dimensions/0/summary.json"
#> [11] "reduced_dimensions/1/content"      "reduced_dimensions/1/summary.json"
#> [13] "row_data/content"                  "row_data/summary.json"            
#> [15] "summary.json"